Move graphql to peerDependencies and devDependencies
Because this fork is meant to publish only the GraphQL schema to npm, only
packages directly used by the schema should be direct `dependencies`. Packages
that are used by the local development server should be in `devDependencies`.
packages of which there should be only one instance (e.g. graphql will error
out if there are multiple instances of graphql) should be in `peerDependencies`.
Pursuant to these guidelines, graphql is moved to both `peerDependencies` and
`devDependencies`. The main impetus is that including it as a direct dependency
causes "multiple graphql instance" error.